修正Controller.php
